+++ examples/hot_potato_game.rs
@@ -92,7 +92,7 @@ impl Handshaking for Player {
         let peer_name = match !conn.side {
             ConnectionSide::Initiator => {
                 // send own PlayerName
-                let own_name = conn.node.name().as_bytes().to_vec();
+                let own_name = self.node().name().as_bytes().to_vec();
                 conn.writer().write_all(&own_name).await?;
 
                 // receive the peer's PlayerName
@@ -106,7 +106,7 @@ impl Handshaking for Player {
                 let peer_name = String::from_utf8_lossy(&buffer[..len]).into_owned();
 
                 // send own PlayerName
-                let own_name = conn.node.name().as_bytes().to_vec();
+                let own_name = self.node().name().as_bytes().to_vec();
                 conn.writer().write_all(&own_name).await?;
 
                 peer_name