+++ src/protocols/handshaking.rs
@@ -27,26 +27,26 @@ where
             while let Some((conn, result_sender)) = from_node_receiver.recv().await {
                 let addr = conn.addr;
 
-                let self_clone2 = self_clone.clone();
+                let node = self_clone.clone();
                 task::spawn(async move {
-                    debug!(parent: conn.node.span(), "handshaking with {} as the {:?}", addr, !conn.side);
+                    debug!(parent: node.node().span(), "handshaking with {} as the {:?}", addr, !conn.side);
                     let result = timeout(
-                        Duration::from_millis(conn.node.config().max_handshake_time_ms),
-                        self_clone2.perform_handshake(conn),
+                        Duration::from_millis(node.node().config().max_handshake_time_ms),
+                        node.perform_handshake(conn),
                     )
                     .await;
 
                     let ret = match result {
                         Ok(Ok(conn)) => {
-                            debug!(parent: self_clone2.node().span(), "succeessfully handshaken with {}", addr);
+                            debug!(parent: node.node().span(), "succeessfully handshaken with {}", addr);
                             Ok(conn)
                         }
                         Ok(Err(e)) => {
-                            error!(parent: self_clone2.node().span(), "handshake with {} failed: {}", addr, e);
+                            error!(parent: node.node().span(), "handshake with {} failed: {}", addr, e);
                             Err(e)
                         }
                         Err(_) => {
-                            error!(parent: self_clone2.node().span(), "handshake with {} timed out", addr);
+                            error!(parent: node.node().span(), "handshake with {} timed out", addr);
                             Err(io::ErrorKind::TimedOut.into())
                         }
                     };